backend/requirements/common.txt
django
djangorestframework
psycopg2
pyyaml 
uritemplate
djangorestframework-simplejwt
drf-spectacular